Bug fix: Support unquoted profile names in cfg file
Fixes #43.  The issue was that building the list of
available profiles vs. retrieving a profile did not go
through the same code path, so there were cases where
an available profile couldn't be retrieved (for example when
the profile name was not quoted).

To fix this, available_profiles and get_config use the same
code to build a mapping of profile name -> config so that any
profile in available_profiles can be used.

Added a test to verify we properly retrieve credentials from
a profile without quotes.