Refactor hooks to precompute handlers to call
When a hierarchical event is emitted, we first do a
lookup to figure out *all* the handlers we need to call
(and in the proper order), then we go about actually calling
the handlers.

We can use this same approach for wildcard events.  One functional
change here is that the event name is kept the same for each
level of the hierarchy.  I think this makes more sense than the
original approach.  So for example, if I emit 'foo.bar.baz', I'd
actually emit 3 events: 'foo.bar.baz', 'foo.bar', 'foo'.  Now,
we calculate the handlers for each of those three events, but logically
we're emitting an event 'foo.bar.baz'.   The tests have been updated
to reflect this.