Allow operations to be renamed with a regex
There's a tradeoff to be made with how general this mechanism
should be vs. what we need now.  What I think is a reasonable
middle ground is:

* Define a new top level 'transformations' key
* Each transformation is keyed by a type, which can be
  anything.

So I added an 'operation-name' transformation that right now just
supports removing part of the name via a regex.  This is because
we only need this specific functionality for now, but it has
potential to grow as needed.  For example::

  {"operation-name": {"replace": ["OriginalText", "ReplacedText"]}
  {"operation-name": {"add": "Suffix"}}

You could also have a new transformation type like:

  {"docs": {"remove-tags": ["examples", "CDATA"]}}