Add validation to verify list members are validated
Part of this change includes providing good error message
for list validations.

If an element in a list fails validation, we get an error
message that does not contain the param.  This is because the
list element failed the validation, not the list param itself.
The element validations are caught and reraised with a more
helpful error message.