Change set_stream_logger to set logger at DEBUG
The issue is that in some situations, we weren't getting debug logs
from botocore or the CLI.  The issue is that by setting the stream
logger to a level of ERROR, we'll never get any debug logs emitted.

This change sets the logger to have a level of DEBUG, but leaves the
associated StreamHandler to have the associated level that's been passed
in.  This should result in no noticeable changes (`--debug` still has
the same result), but this has the benefit that you can add addtional
handlers that record at the DEBUG level (say in a plugin).

This particular situation was noticed because when a test fails, nose
normally captures all the log messages, which is helpful in diagnosing
what went wrong.  However, because clidrive calls set_stream_logger with
a level of ERROR, nose's log handlers can't capture any debug logs.