Apply operation renames before pagination configs
This allows you to refer to operations by their transformed names
instead of their names from the model.

This was specifically for cloudfront, where we strip off the API version
from the operation names, but want to apply the same pagination configs
defined in cloudfront.extra.json.  Otherwise we have to update the
pagination configs every time we update the model.