Remove cli_name from .extra files
This logic can live in the argrename CLI customization.

The only one I didn't pull out was the sns one for endpoint,
because that also has a py_name attribute as well.

I think we might be able to move that out, but I'll have to double
check, and if so, it will be in a separate commit/PR.