Revert "Merge branch 'expect-100' into develop"
The expect-100 changes do not work in python2.6.

The internal structure of HTTPConnection is different
in python2.6 so this needs to be reworked.

I'm still fairly confident that this can be done in python2.6,
it just needs additional research and time.

This reverts commit 09f41601be41f98b616aab4c240385b75be0ea9f, reversing
changes made to ead5601691fe85116ab09b41c65f48484f8039c5.