Move auth_path handling to the request.
This fixes the previous issue where this was not fixed
previously by checking if the auth_path was not None.
By moving the handling of the auth_path onto the request
itself, this makes is easier to ensure that the auth_path
manipulation only happens on a per request basis.