Revert "Merge branch 'per-client-events' into develop"
This reverts commit b3647b69ce00ffa4a4edee9ff1725ce2a810d065, reversing
changes made to 838a9dd836045c016952f5f4b9a0e8093015d86d.

This is causing test failures in the integration tests.
It appears to happen in the deepcopy code when we encounter
a partial object.

We also don't encounter a partial object unless we also use
a session to call .get_service(), which registers additional handlers.
This will require some investigation so I'm reverting in the meantime.