Rename .api.json to .normal.json
We should match the existing convention from the
upstream sources which use .api.json and .normal.json.

This is also the suffix that other SDKs use (.normal.json),
so it causes confusion with botocore being different.  We
should be consistent and have .normal.json be the complete
model with docs (which is what we use) and have .api.json
be the model without docs.

This isn't exposed directly to customers so this is
a backwards compatible change.

Note that there are no functional changes here, these are the
same models just renamed.