Use GMT for timestamp format
This changes our date serialization to
use an IMF-fixdate timestamp in headers for the rest serializers.

RFC 7231 describes the preferred date-time format in HTTP messages as an
IMF-fixdate timestamp. This timestamp format is a subset of the format
described in RFC 5322 except the format always used a fixed timezone offset of
"GMT". S3 also uses this timestamp format in their documentation and in the
response headers they return.