Adding an enum attribute and StringShape.
This commit updates the shape builders to utilize a new concrete shape
type, StringShape. StringShape exposes a property that makes it easier
to inspect the enum attribute of a string shape. StringShape is the only
concrete shape type to expose this property becase, at present, the only
shapes that have the `enum` property are string shapes.

In addition to this feature, the unused SHAPE_CLASSES constant has been
deleted from ServiceModel.