add tests to MANIFEST.in
When packaging botocore it's nice to be able to run the tests from the
source distribution to ensure that the packaging has been done
correctly.  This includes the tests directoy in the sdist without
installing it to the system.  Piggy-backed is a change to the docs
inclusion that uses graft rather than recursive-include.