Update serializers/parsers with latest protocol tests
There were a couple of edge cases we were not handling properly.
One of the main changes added was to support sequences as
values in a query string.  This means that something like
"{"a": ['x', 'y']}" will now be serialized as
"a=x&a=y".

Also handle non flattened list with locationName in query protocol.