Reset HTTP connection state when connection is closed
A connection instance can be reused even if the server
has closed the associated socket.

When this happens, urllib3 will ".close()" the connection
which also closes and deletes the socket associated with
the connection.

However, the AWSHTTPConnection *object* is still reused
in the underlying connection pool.  The HTTP client will
notice that associated socket is None and will create
a new socket thereby establishing a new connection to the
AWS endpoint.

The bug on our end was that we were not properly resetting
our own instance state variables we use to keep track of
if we've sent the expect header, if we've seen a response,
etc.  We were assuming that a connection instance is removed
from the connection pool when the connection and underlying
socket is closed.

As a result, state from the previous (and generally unrelated)
response was leaking into the current response, specifically
in the scenario when the service responds with some other
status besides the "100 continue" response (i.e a fast fail).
This ended up with two things happening:

* We were reusing the status code of the previous response
* Because we thought we've already seen a response, we were
  not reading the HTTP response headers.

Consequently, the HTTP headers were being sent to our
response parser, which would throw exceptions because it
was not receiving data in the format it was expecting.

We now ensure that we also reset our own state when the
parent class resets its state, which happens in the
".close()" method.

In addition to the associated test, I've also manually
verified this fix by writing a small http server that
returns the same sequence of HTTP responses that trigger
this error.