Profile should override env var creds
Unless the profile and creds are both specified
as env vars.

In order to fix this regression, I had to make a
small change to session.  This makes the 'profile'
config var less special cased than it was previously.

Now, we always check the _session_instance_vars for
a profile, and simple use _profile as a placeholder
for the cached lookup value of 'profile'.

The only place this gets weird is we need to scope
down our call to ".get_config_variable" in the
credential provider creation to only check the
session instance variables for a profile to determine
whether or not we need to add the EnvProvider to our
lookup chain.

One of the reasons this regressed is we didn't have
sufficient functional tests to cover this behavior.
I've added functional tests that ensure that regardless
of internals, we won't regress on this in the future.

There will also be a pull request to the AWS CLI
to verify this behavior to ensure the CLI doesn't regress
on this behavior.

Fixes #663.