Initial commit of acceptance tests
These add the basic smoke tests written in
gherkin.  These are the same shared tests across
the AWS SDKs.

I think the eventual plan is to remove most the test_smoke.py
tests in favor of these tests.

For now, I've added the tests so we can run the manually until
we're ready to make the switch.

The tests are all passing.