Add support for default client config on session
Exposed a set_default_client_config() and get_default_client_config() that
will be used as the config for when clients are created. If a config is
also passed to create_client() the config will be merged into the default
client config and only override default client config values with nondefault
values provided to the client config. Also added a merge() method on client
config to help facilitate the merging of configs.